+++ cmd/influx_inspect/help/help.go
@@ -31,6 +31,7 @@ Usage: influx_inspect [[command] [arguments]]
 
 The commands are:
 
+    dumptsi              dumps low-level details about tsi1 files.
     dumptsm              dumps low-level details about tsm1 files.
     export               exports raw data from a shard to line protocol
     help                 display this help message