+++ tsdb/index/inmem/inmem.go
@@ -156,10 +156,11 @@ func (i *Index) CreateSeriesIfNotExists(shardID uint64, key, name []byte, tags m
 	m := i.CreateMeasurementIndexIfNotExists(name)
 
 	i.mu.Lock()
+	defer i.mu.Unlock()
+
 	// Check for the series again under a write lock
 	ss = i.series[string(key)]
 	if ss != nil {
-		i.mu.Unlock()
 		ss.AssignShard(shardID)
 		return nil
 	}
@@ -167,7 +168,6 @@ func (i *Index) CreateSeriesIfNotExists(shardID uint64, key, name []byte, tags m
 	// Verify that the series will not exceed limit.
 	if !ignoreLimits {
 		if max := opt.Config.MaxSeriesPerDatabase; max > 0 && len(i.series)+1 > max {
-			i.mu.Unlock()
 			return errMaxSeriesPerDatabaseExceeded
 		}
 	}
@@ -186,7 +186,6 @@ func (i *Index) CreateSeriesIfNotExists(shardID uint64, key, name []byte, tags m
 
 	// Add the series to the series sketch.
 	i.seriesSketch.Add(key)
-	i.mu.Unlock()
 
 	return nil
 }
@@ -225,9 +224,9 @@ func (i *Index) CreateMeasurementIndexIfNotExists(name []byte) *Measurement {
 // HasTagKey returns true if tag key exists.
 func (i *Index) HasTagKey(name, key []byte) (bool, error) {
 	i.mu.RLock()
-	defer i.mu.RUnlock()
-
 	mm := i.measurements[string(name)]
+	i.mu.RUnlock()
+
 	if mm == nil {
 		return false, nil
 	}
@@ -249,9 +248,9 @@ func (i *Index) HasTagValue(name, key, value []byte) bool {
 // TagValueN returns the cardinality of a tag value.
 func (i *Index) TagValueN(name, key []byte) int {
 	i.mu.RLock()
-	defer i.mu.RUnlock()
-
 	mm := i.measurements[string(name)]
+	i.mu.RUnlock()
+
 	if mm == nil {
 		return 0
 	}
@@ -261,9 +260,9 @@ func (i *Index) TagValueN(name, key []byte) int {
 // MeasurementTagKeysByExpr returns an ordered set of tag keys filtered by an expression.
 func (i *Index) MeasurementTagKeysByExpr(name []byte, expr influxql.Expr) (map[string]struct{}, error) {
 	i.mu.RLock()
-	defer i.mu.RUnlock()
-
 	mm := i.measurements[string(name)]
+	i.mu.RUnlock()
+
 	if mm == nil {
 		return nil, nil
 	}
@@ -295,9 +294,9 @@ func (i *Index) ForEachMeasurementTagKey(name []byte, fn func(key []byte) error)
 // TagKeyCardinality returns the number of values for a measurement/tag key.
 func (i *Index) TagKeyCardinality(name, key []byte) int {
 	i.mu.RLock()
-	defer i.mu.RUnlock()
-
 	mm := i.measurements[string(name)]
+	i.mu.RUnlock()
+
 	if mm == nil {
 		return 0
 	}
@@ -307,9 +306,9 @@ func (i *Index) TagKeyCardinality(name, key []byte) int {
 // TagsForSeries returns the tag map for the passed in series
 func (i *Index) TagsForSeries(key string) (models.Tags, error) {
 	i.mu.RLock()
-	defer i.mu.RUnlock()
-
 	ss := i.series[key]
+	i.mu.RUnlock()
+
 	if ss == nil {
 		return nil, nil
 	}
@@ -521,10 +520,11 @@ func (i *Index) DropSeries(key []byte) error {
 	}
 
 	i.mu.Lock()
+	defer i.mu.Unlock()
+
 	k := string(key)
 	series := i.series[k]
 	if series == nil {
-		i.mu.Unlock()
 		return nil
 	}
 
@@ -541,7 +541,6 @@ func (i *Index) DropSeries(key []byte) error {
 	if !series.Measurement().HasSeries() {
 		i.dropMeasurement(series.Measurement().Name)
 	}
-	i.mu.Unlock()
 
 	return nil
 }